# Add `ops.nn.moments` and speed-up normalization layers