EDITED:
This PR adds `ops.nn.moments` and improves some normalization layers by fast mean and variance computation.
This PR also addresses the overflow & underflow issue when the input tensor is float16.

There are two approaches to compute variance:
- $Var = E[|x - E[x]|^2])$ (stable but slower)
- $Var = E[|x|^2] - |E[x]|^2$ (fast but unstable)

|backend|layer|manual implementation (before this PR)|`ops.nn.moments` (fast but unstable, this PR)|`ops.nn.moments` (stable but slower)|
|-|-|-|-|-|
|tensorflow|BatchNormalization|58ms|58ms|69ms|
|jax|BatchNormalization|63ms|63ms|75ms|
|torch|BatchNormalization|73ms|72ms|74ms|
|tensorflow|GroupNormalization|89ms|61ms|73ms|
|jax|GroupNormalization|96ms|72ms|83ms|
|torch|GroupNormalization|72ms|74ms|76ms|
|tensorflow|LayerNormalization|52ms|47ms|48ms|
|jax|LayerNormalization|68ms|59ms|60ms|
|torch|LayerNormalization|88ms|90ms|91ms|

References:
- tensorflow: `tf.nn.moments` using stable but slower version ([link](https://github.com/tensorflow/tensorflow/blob/3d1802023778a164d35c79536990b35b701e8018/tensorflow/python/ops/nn_impl.py#L1264C5-L1268C25))
- flax: defaults to `use_fast_variance` ([link](https://github.com/google/flax/blob/ca3ea06f78834137dfb49dc6c1a0c26fb962003a/flax/linen/normalization.py#L108-L120))
- torch: should be stable but slower version ([link](https://github.com/pytorch/pytorch/blob/48e6ffbe308e915b67c5b4f9532f794d6706c903/aten/src/ATen/native/cpu/batch_norm_kernel.cpp#L200-L209))