@fchollet 2023-09-12
Thanks for the PR, this is great! What performance changes did you observe in e.g. BatchNormalization when using `moments` in different backends compared to the existing manual implementation?