@james77777778 2023-09-13
> Thanks for the PR, this is great! What performance changes did you observe in e.g. BatchNormalization when using `moments` in different backends compared to the existing manual implementation?

I have reordered the operations to get a significant speed-up. The key is to downsize the tensor before performing element-wise operations (such as `jnp.subtract`)
```python
# actually, tf.nn.moments could be faster...

# original
variance = jnp.mean(jnp.square(x - jax.lax.stop_gradient(mean)), axis=axes, keepdims=True)

# faster version
variance = jnp.mean(jnp.square(x), axis=axes, keepdims=True) - jnp.square(
    jax.lax.stop_gradient(mean)
)
```
I have also updated the Normalization layers. I observed better performance in TF & JAX but torch showed similar performance.

Benchmark script:
```python
from keras_core import layers
from keras_core import models
from keras_core import ops

x_train = ops.random.uniform(shape=(1024, 224, 224, 3))
y_train = ops.random.uniform(shape=(1024, 224, 224, 3))

# layers.BatchNormalization
# layers.GroupNormalization
# layers.LayerNormalization
normalization_cls = layers.BatchNormalization
normalization_args = {}
if normalization_cls is layers.GroupNormalization:
    normalization_args = {"groups": 3}

model = models.Sequential(
    [
        layers.InputLayer(shape=(224, 224, 3)),
        normalization_cls(**normalization_args),
        normalization_cls(**normalization_args),
        normalization_cls(**normalization_args),
    ]
)
model.compile(loss="mse", optimizer="adam")
model.fit(x_train, y_train, batch_size=128, epochs=3)

```

Results (with 1080 8gb card):

|backend|layer|current implementation|using `ops.nn.moments`|notes|
|-|-|-|-|-|
|tensorflow|BatchNormalization|58ms|58ms|fair|
|jax|BatchNormalization|63ms|63ms|fair|
|torch|BatchNormalization|73ms|72ms|fair|
|tensorflow|GroupNormalization|89ms|61ms|⬇️|
|jax|GroupNormalization|96ms|72ms|⬇️|
|torch|GroupNormalization|72ms|74ms|fair|
|tensorflow|LayerNormalization|52ms|47ms|⬇️|
|jax|LayerNormalization|68ms|59ms|⬇️|
|torch|LayerNormalization|88ms|90ms|fair|