@fchollet 2023-09-13
Thanks for the analysis -- what does "current implementation" refer to? The implementation in this PR?