@james77777778 2023-09-14
> Thanks for the analysis -- what does "current implementation" refer to? The implementation in this PR?

Sorry for the confusion. I have updated the table as as follows:

|backend|layer|manually implementation (before this PR)|`ops.nn.moments` (fast but unstable variance computation)|`ops.nn.moments` (slow but stable variance computation)|
|-|-|-|-|-|
|tensorflow|BatchNormalization|58ms|58ms|69ms|
|jax|BatchNormalization|63ms|63ms|TBD|
|torch|BatchNormalization|73ms|72ms|TBD|
|tensorflow|GroupNormalization|89ms|61ms|73ms|
|jax|GroupNormalization|96ms|72ms|TBD|
|torch|GroupNormalization|72ms|74ms|TBD|
|tensorflow|LayerNormalization|52ms|47ms|48ms|
|jax|LayerNormalization|68ms|59ms|TBD|
|torch|LayerNormalization|88ms|90ms|TBD|