@fchollet 2023-09-14
My take is that until we see reports of users running into stability issues, then the fast implementation should be fine. The fact that Flax defaults to it is evidence that there's little issue. I did a quick search within the google codebase and found only a couple of usages of `use_fast_variance=False`, among thousands of usages of Flax normalization layers (they aren't commented, so unclear why they went with `False`). So it seems that in practice the problem doesn't really surface.