@fchollet 2023-09-14
The code looks good! Do you want to include the normalization layer changes in this PR, or merge this PR first and then create another one?