@james77777778 2023-09-14
> My take is that until we see reports of users running into stability issues, then the fast implementation should be fine. The fact that Flax defaults to it is evidence that there's little issue. I did a quick search within the google codebase and found only a couple of usages of `use_fast_variance=False`, among thousands of usages of Flax normalization layers (they aren't commented, so unclear why they went with `False`). So it seems that in practice the problem doesn't really surface.

Thanks for the valuable insights about the usage of Flax.
After a lot of searching, I can only find this comment defending the numerically stable computation of variance (without evidence?) https://github.com/tensorflow/tensorflow/pull/4198#issuecomment-244824858

> The code looks good! Do you want to include the normalization layer changes in this PR, or merge this PR first and then create another one?

I think the changes are already in this PR. Please let me know if I missed anything.