@james77777778 2023-09-14
Hi @fchollet 
This PR should be ready. We now have fast mean & variance computation using `ops.nn.moments`, and it is applied to BatchNormalization, GroupNormalization and LayerNormalization to achieve some speed-ups.

Addtionally, this PR addresses the overflow and underflow issue that occur when the input is float16.  (I encountered this before when using GroupNormalization with mixed_float16)
Credits to @fsx950223 (https://github.com/tensorflow/tensorflow/pull/52217)